+++ db/migrate/20260715072550_create_instagram_connect_messages.rb
@@ -1,4 +1,4 @@
-class CreateInstagramConnectMessages < ActiveRecord::Migration[<%= ActiveRecord::Migration.current_version %>]
+class CreateInstagramConnectMessages < ActiveRecord::Migration[7.1]
   def change
     create_table :instagram_connect_messages do |t|
       t.bigint :conversation_id, null: false