+++ db/migrate/20260715072552_create_instagram_connect_comments.rb
@@ -1,4 +1,4 @@
-class CreateInstagramConnectComments < ActiveRecord::Migration[<%= ActiveRecord::Migration.current_version %>]
+class CreateInstagramConnectComments < ActiveRecord::Migration[7.1]
   def change
     create_table :instagram_connect_comments do |t|
       t.bigint :account_id, null: false